third_party/WebKit/LayoutTests/css3/device-adapt/viewport-initial-height-zero-expected.txt

ALERT: viewport size 200x100 scale 1 with limits [1, -1] and userScalable true
